Widen the joystick report to 32 buttons
Eight buttons cannot represent a two-Joy-Con pad: four face, four D-pad,
four shoulders, two stick clicks, start/select, home and capture already
exceed it, and a real D-pad deserves its own slots rather than being folded
into the face cluster.

Widen the flags type to uint32_t and the descriptor's button REPORT_COUNT to
32 together so the report body stays self-consistent; every send path
already derives its length from sizeof(report), so nothing else needs
touching. The press-count table now sizes itself from
ZMK_HID_JOYSTICK_NUM_BUTTONS rather than a hardcoded 8.

+++ include/zmk/hid-io/joystick.h
@@ -6,5 +6,11 @@
 
 #pragma once
 
-typedef uint8_t zmk_joystick_button_flags_t;
+/* 32 buttons: a two-Joy-Con pad needs ~14 (4 face + dpad + 4
+ * shoulders + stick clicks + start/select), which does not fit
+ * in 8. Widening the flags type and the descriptor's button
+ * count together keeps the report body self-consistent. 32 leaves
+ * room for a real D-pad alongside the face cluster and both
+ * shoulder pairs, and matches what many PC pads report. */
+typedef uint32_t zmk_joystick_button_flags_t;
 typedef uint16_t zmk_joystick_button_t;
